+++ mobile/lib/screens/insights/insights_page.dart
@@ -1,11 +1,11 @@
 import 'package:app/constants/config.dart';
 import 'package:app/models/place_details.dart';
+import 'package:app/screens/insights/insights_tab.dart';
 import 'package:app/widgets/custom_widgets.dart';
-import 'package:app/widgets/insights_tab.dart';
 import 'package:flutter/material.dart';
 
-import '../models/enum_constants.dart';
-import '../themes/light_theme.dart';
+import '../../models/enum_constants.dart';
+import 'insights_widgets.dart';
 
 class InsightsPage extends StatefulWidget {
   final PlaceDetails placeDetails;
@@ -58,8 +58,8 @@ class _InsightsPageState extends State<InsightsPage>
                       }
                     },
                     tabs: <Widget>[
-                      tabButton(text: 'Day'),
-                      tabButton(text: 'Week'),
+                      InsightsTabButton(text: 'Day', frequency: frequency),
+                      InsightsTabButton(text: 'Week', frequency: frequency),
                     ]),
               ),
             ),
@@ -88,32 +88,5 @@ class _InsightsPageState extends State<InsightsPage>
   void initState() {
     super.initState();
     _tabController = TabController(length: 2, vsync: this);
-  }
-
-  Widget tabButton({required String text}) {
-    return Container(
-      constraints:
-          const BoxConstraints(minWidth: double.infinity, maxHeight: 32),
-      decoration: BoxDecoration(
-          color: text.toLowerCase() == 'day'
-              ? frequency == Frequency.hourly
-                  ? Config.appColorBlue
-                  : Colors.white
-              : frequency == Frequency.hourly
-                  ? Colors.white
-                  : Config.appColorBlue,
-          borderRadius: const BorderRadius.all(Radius.circular(4.0))),
-      child: Tab(
-          child: Text(text,
-              style: CustomTextStyle.button1(context)?.copyWith(
-                color: text.toLowerCase() == 'day'
-                    ? frequency == Frequency.hourly
-                        ? Colors.white
-                        : Config.appColorBlue
-                    : frequency == Frequency.hourly
-                        ? Config.appColorBlue
-                        : Colors.white,
-              ))),
-    );
   }
 }
