+++ mobile/lib/screens/profile/profile_edit_page.dart
@@ -291,7 +291,7 @@ class _ProfileEditPageState extends State<ProfileEditPage> {
         automaticallyImplyLeading: false,
         title: Row(
           children: [
-            backButton(context),
+            const AppBackButton(),
             const Spacer(),
             Text(
               'Edit Profile',
